# [core] Optimize MAC address formatting to eliminate sprintf dependency